requirements.txt
python-telegram-bot==13.15
requests==2.32.3
python-dotenv==1.0.1
urllib3==1.26.18
google_cloud_dialogflow==2.41.1
vk_api==11.9.9